requirements.txt
Django
click
colorama
django-bootstrap4
django-crispy-forms
django-filter
autologging
django-vite
django-tables2
django-cors-headers
fabric
django-recaptcha
lxml
numpy
pandas
Pillow
django-debug-toolbar
plotly
pyparsing
pypdf
django-webpack-loader
reportlab
django-allauth
django-extensions
django-dynamic-breadcrumbs
django-autocomplete-light
crispy-bootstrap4
django-storages
reportlab
requests
sourcery
statsmodels
stone
strsimpy
svglib
tablib
weasyprint
Whoosh
xarray
xhtml2pdf
xlrd
XlsxWriter
xlwt
redis
pilkit
django-redis
celery
django-celery-beat
django-celery-results
Faker
pylint
pylint-django
dj-stripe
django-meta
django-mail-templated
celery-progress
python-decouple
channels
djangorestframework
django-grappelli